+++ README.md
@@ -35,9 +35,9 @@ fun main() = object : KiloEngineGame {
     init {
         Resources.resourceGetter += javaClass::getResourceAsStream
     }
-    
+
     override val title get() = "My Game"
-    
+
     override fun loadInitScene() {
         sceneLoader.loadScene("/my_game/my_scene.scene.json5", scene)
     }
@@ -55,13 +55,13 @@ public class MyGame extends KiloEngineGame {
         Resources.addResourceGetter(MyGame.class::getResourceAsStream);
         new MyGame().main();
     }
-    
+
     @Override
     @NotNull
     public String getTitle() {
         return "My Game";
     }
-    
+
     @Override
     public void loadInitScene() {
         getSceneLoader().loadScene("/my_game/my_scene.scene.json5", getScene());
@@ -77,25 +77,16 @@ A game in Kilo Engine is made up of "scenes" (well actually just one at this poi
 
 Scenes can be constructed purely through using code, but it is preferred to specify the initial data of scenes through `.scene.json5` files. You can load scenes from Json5 files using `sceneLoader#loadScene`.
 
-Scenes, however, have a bit more to them than just entities. Specifically, they can also have a skybox and a sun. The skybox is a cubemap defining the sides. However, a skybox is optional, and although one will be provided by default - you can remove it. Please do note, however, that a skybox should almost *always* be specified for outdoor scenes, as the color buffer is never cleared, so the sky without a skybox will just be the previous frame. The sun position indicates a position for the sun. The sun will be a directional light from its position to `(0, 0, 0)`. If the sun's position is not specified, no lighting will be automatically applied.
+Scenes, however, have a bit more to them than just entities. Specifically, they can also have a skybox and a sun. The skybox is a cubemap defining the sides, a solid color, or nothing. If no skybox is specified, it will default to no skybox. Please do note, however, that a skybox should almost *always* be specified for outdoor scenes, as the color buffer is never cleared, so the sky without a skybox will just be the previous frame. The sun position indicates a position for the sun. The sun will be a directional light from its position to `(0, 0, 0)`. If the sun's position is not specified, no lighting will be automatically applied.
 
 Scene Json5 syntax is simple. Here's a brief example:
 
-<details>
+<details open>
 <summary>Sample Json5 scene</summary>
 
 ```json5
 {
-  // if you want to remove the skybox:
-  skybox: null,
-  
-  // otherwise:
-  skybox: { // Skybox is optional. See above.
-    
-    // if you want to use a single texture, use the following:
-    all: "/example/skybox/down.png",
-    
-    // otherwise, use this for multiple textures:
+  skybox: { // A cubemap skybox
     base: "/example/skybox", // *Optional* base path for the paths in this object
     down: "down.png", // Because `base` is specified, this will load from `/example/skybox/down.png`.
     up: "up.png",
@@ -104,6 +95,7 @@ Scene Json5 syntax is simple. Here's a brief example:
     negativeX: "negativeX.png",
     positiveX: "positiveX.png"
   },
+  skybox: [0.75, 0.25, 0.25], // A light red solid color skybox
   sunPosition: [-12.9, 30, 17.1], // Sun position is optional. See above.
   entities: [
     // What's a level without some entities, eh?
@@ -153,7 +145,7 @@ So you want to see how this is done in code? Well here you go:
 <summary>The above example, but in Kotlin</summary>
 
 ```kotlin
-scene.skybox = SkyboxTextures.relative(
+scene.skybox = Skybox.Cubemap.relative(
     "/example/skybox",
     "down.png",
     "up.png",
@@ -162,12 +154,13 @@ scene.skybox = SkyboxTextures.relative(
     "negativeX.png",
     "positiveX.png"
 )
+scene.skybox = Skybox.SolidColor(0.75f, 0.25f, 0.25f)
 scene.sunPosition = Vector3f(-12.9f, 30f, 17.1f)
 Entity(scene, DVector3()).apply {
     body.setKinematic()
     val model = MeshComponent(this, sceneLoader.loadObj("/example/example.obj")).model
     MeshRendererComponent(this)
-    MeshColliderComponent(this, CollisionModel(model, mapOf( 
+    MeshColliderComponent(this, CollisionModel(model, mapOf(
         model.materials["Brick_Antique_01"]!! to CollisionTypes.WALL,
         model.materials["Brick_Basket"]!! to CollisionTypes.FLOOR,
         model.materials["Death_Plane"]!! to CollisionTypes.DEATH
@@ -186,38 +179,35 @@ Entity(scene, DVector3(0.0, 1.4, -5.0)).apply {
 <summary>The above example, but in Java</summary>
 
 ```java
-class Test {
-    void test() {
-        final Scene scene = getScene();
-        scene.setSkybox(SkyboxTextures.relative(
-            "/example/skybox",
-            "down.png",
-            "up.png",
-            "negativeZ.png",
-            "positiveZ.png",
-            "negativeX.png",
-            "positiveX.png"
-        ));
-        scene.setSunPosition(new Vector3f(-12.9f, 30, 17.1f));
-        
-        final Entity levelMesh = new Entity(scene, new DVector3());
-        levelMesh.getBody().setKinematic();
-        final Model model = new MeshComponent(
-            levelMesh, getSceneLoader().loadObj("/example/example.obj")
-        ).getModel();
-        new MeshRendererComponent(levelMesh);
-        new MeshColliderComponent(levelMesh, new CollisionModel(model, Map.of(
-            model.getMaterials().get("Brick_Antique_01"), CollisionTypes.WALL,
-            model.getMaterials().get("Brick_Basket"), CollisionTypes.FLOOR,
-            model.getMaterials().get("Death_Plane"), CollisionTypes.DEATH
-        )));
-        
-        final Entity player = new Entity(scene, new DVector3(0, 1.4, -5));
-        new CapsuleColliderComponent(player, 0.4, 1);
-        new PlayerComponent(player, new Vector2f());
-        new CameraComponent(player, new DVector3(0.0, 0.7, 0.0));
-    }
-}
+final Scene scene = getScene();
+scene.setSkybox(Skybox.Cubemap.relative(
+    "/example/skybox",
+    "down.png",
+    "up.png",
+    "negativeZ.png",
+    "positiveZ.png",
+    "negativeX.png",
+    "positiveX.png"
+));
+scene.setSkybox(new Skybox.SolidColor(0.75f, 0.25f, 0.25f));
+scene.setSunPosition(new Vector3f(-12.9f, 30, 17.1f));
+
+final Entity levelMesh = new Entity(scene, new DVector3());
+levelMesh.getBody().setKinematic();
+final Model model = new MeshComponent(
+    levelMesh, getSceneLoader().loadObj("/example/example.obj")
+).getModel();
+new MeshRendererComponent(levelMesh);
+new MeshColliderComponent(levelMesh, new CollisionModel(model, Map.of(
+    model.getMaterials().get("Brick_Antique_01"), CollisionTypes.WALL,
+    model.getMaterials().get("Brick_Basket"), CollisionTypes.FLOOR,
+    model.getMaterials().get("Death_Plane"), CollisionTypes.DEATH
+)));
+
+final Entity player = new Entity(scene, new DVector3(0, 1.4, -5));
+new CapsuleColliderComponent(player, 0.4, 1);
+new PlayerComponent(player, new Vector2f());
+new CameraComponent(player, new DVector3(0.0, 0.7, 0.0));
 ```
 
 </details>
