+++ src/main/kotlin/io/github/gaming32/kiloengine/Scene.kt
@@ -20,7 +20,7 @@ class Scene {
     @PublishedApi
     internal val bodyToEntity = mutableMapOf<DBody, Entity>()
 
-    var skybox: SkyboxTextures? = SkyboxTextures.DEFAULT
+    var skybox: Skybox = Skybox.None
     var sunPosition: Vector3f? = null
 
     private var eventsDirty = false
