+++ src/main/kotlin/io/github/gaming32/kiloengine/loader/SceneLoaderImpl.kt
@@ -5,7 +5,7 @@ import com.google.gson.JsonParser
 import io.github.gaming32.gson5.Gson5Reader
 import io.github.gaming32.kiloengine.ResourceGetter
 import io.github.gaming32.kiloengine.Scene
-import io.github.gaming32.kiloengine.SkyboxTextures
+import io.github.gaming32.kiloengine.Skybox
 import io.github.gaming32.kiloengine.entity.ComponentRegistry
 import io.github.gaming32.kiloengine.entity.Entity
 import io.github.gaming32.kiloengine.model.Material
@@ -104,15 +104,7 @@ class SceneLoaderImpl(private val resourceGetter: () -> ResourceGetter) : SceneL
 
     override fun loadScene(name: String, scene: Scene) = textResource(name) { inp, _ ->
         val json = JsonParser.parseReader(Gson5Reader(JsonReader.json5(inp))).asJsonObject
-        json["skybox"]?.let {
-            if (it.isJsonObject) {
-                scene.skybox = SkyboxTextures.fromJson(it.asJsonObject)
-            } else if (it.isJsonNull) {
-                scene.skybox = null
-            } else {
-                throw IllegalArgumentException("Expected null or object for skybox, not $it")
-            }
-        }
+        json["skybox"]?.let { scene.skybox = Skybox.fromJson(it) }
         json["sunPosition"]?.asJsonArray?.let { scene.sunPosition = it.toVector3f() }
         json["gravity"]?.asJsonArray?.let { scene.world.setGravity(it.toDVector3()) }
         json["entities"]?.asJsonArray?.forEach { entityData ->
